refactor(android): shared 하위의 DTO를 사용하는 것으로 리팩터링
