--- server/src/main/kotlin/good/space/runnershi/global/auth/RefreshTokenRequest.kt
@@ -1,5 +1,0 @@
-package good.space.runnershi.global.auth
-
-data class RefreshTokenRequest(
-    val refreshToken: String
-)
