--- shared/src/commonMain/kotlin/good/space/runnershi/model/dto/LoginRequest.kt
@@ -1,15 +1,0 @@
-package good.space.runnershi.model.dto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class LoginRequest(
-    val email: String,
-    val password: String
-)
-
-@Serializable
-data class LoginResponse(
-    val accessToken: String,
-    val refreshToken: String
-)
