--- shared/src/commonMain/kotlin/good/space/runnershi/model/dto/RefreshTokenRequest.kt
@@ -1,9 +1,0 @@
-package good.space.runnershi.model.dto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RefreshTokenRequest(
-    val refreshToken: String
-)
-
