--- shared/src/commonMain/kotlin/good/space/runnershi/model/dto/SignUpRequest.kt
@@ -1,11 +1,0 @@
-package good.space.runnershi.model.dto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SignUpRequest(
-    val email: String,
-    val password: String,
-    val nickname: String
-)
-
