--- shared/src/commonMain/kotlin/good/space/runnershi/model/dto/TokenResponse.kt
@@ -1,10 +1,0 @@
-package good.space.runnershi.model.dto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class TokenResponse(
-    val accessToken: String,
-    val refreshToken: String
-)
-
