+++ shared/src/commonMain/kotlin/good/space/runnershi/model/dto/auth/AuthRequestDto.kt
@@ -1,0 +1,21 @@
+package good.space.runnershi.model.dto.auth
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SignUpRequest(
+    val email: String,
+    val password: String,
+    val name: String
+)
+
+@Serializable
+data class LoginRequest(
+    val email: String,
+    val password: String
+)
+
+@Serializable
+data class TokenRefreshRequest(
+    val refreshToken: String
+)
