+++ shared/src/commonMain/kotlin/good/space/runnershi/model/dto/auth/AuthResponseDto.kt
@@ -1,0 +1,20 @@
+package good.space.runnershi.model.dto.auth
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class LoginResponse(
+    val accessToken: String,
+    val refreshToken: String
+)
+
+@Serializable
+data class TokenResponse(
+    val accessToken: String,
+    val refreshToken: String
+)
+
+@Serializable
+data class TokenRefreshResponse(
+    val accessToken: String
+)
